agregar cifrado

+++ src/main/java/net/datatecsolution/admin_tools/controlador/CtlOrdenVenta.java
@@ -776,21 +776,20 @@ public void calcularTotales(){
 				//detalle.setImpuesto(impuesto2.setScale(2, BigDecimal.ROUND_HALF_EVEN));
 				detalle.setTotal(totalItem.setScale(2, BigDecimal.ROUND_HALF_EVEN));
 
-				//se establece el total e impuesto en el vista
-				this.view.getTxtTotal().setText(""+myFactura.getTotal().setScale(2, BigDecimal.ROUND_HALF_EVEN));
-				this.view.getTxtImpuesto().setText(""+myFactura.getTotalImpuesto().setScale(2, BigDecimal.ROUND_HALF_EVEN));
-				this.view.getTxtImpuesto18().setText(""+myFactura.getTotalImpuesto18().setScale(2, BigDecimal.ROUND_HALF_EVEN));
-				this.view.getTxtSubtotal().setText(""+myFactura.getSubTotal().setScale(2, BigDecimal.ROUND_HALF_EVEN));
-				this.view.getTxtDescuento().setText(""+myFactura.getTotalDescuento().setScale(2, BigDecimal.ROUND_HALF_EVEN));
-
-				view.getModeloTabla().fireTableDataChanged();
-				this.selectRowInset();
-
-				view.getTxtBuscar().requestFocusInWindow();
-
 
 				//this.view.getModelo().fireTableDataChanged();
 			}//fin del if
+		//se establece el total e impuesto en el vista
+		this.view.getTxtTotal().setText(""+myFactura.getTotal().setScale(2, BigDecimal.ROUND_HALF_EVEN));
+		this.view.getTxtImpuesto().setText(""+myFactura.getTotalImpuesto().setScale(2, BigDecimal.ROUND_HALF_EVEN));
+		this.view.getTxtImpuesto18().setText(""+myFactura.getTotalImpuesto18().setScale(2, BigDecimal.ROUND_HALF_EVEN));
+		this.view.getTxtSubtotal().setText(""+myFactura.getSubTotal().setScale(2, BigDecimal.ROUND_HALF_EVEN));
+		this.view.getTxtDescuento().setText(""+myFactura.getTotalDescuento().setScale(2, BigDecimal.ROUND_HALF_EVEN));
+
+		view.getModeloTabla().fireTableDataChanged();
+		this.selectRowInset();
+
+		view.getTxtBuscar().requestFocusInWindow();
 
 	}//fin del for
 	}
@@ -1270,6 +1269,7 @@ public void calcularTotales(){
 						if(filaPulsada>=0){
 							 this.view.getModeloTabla().eliminarDetalle(filaPulsada);
 							 this.calcularTotales();
+							 //this.setEmptyView();
 						 }
 						break;
 						
@@ -1584,38 +1584,6 @@ public void calcularTotales(){
 	                //view.getToolkit().beep();
 	        }
 		}
-		/*
-	//que no se la fecha de arriba y abajo
-	if(e.getKeyCode()!=KeyEvent.VK_DOWN && e.getKeyCode()!= KeyEvent.VK_UP && e.getKeyCode()!= KeyEvent.VK_ENTER)
-		
-		//se comprueba que hay algo que buscar
-		if(e.getComponent()==this.view.getTxtBuscar()&&view.getTxtBuscar().getText().trim().length()!=0){
-			
-			myArticuloDao=new ArticuloDao(conexion);
-			//JOptionPane.showMessageDialog(view, "2");
-			//JOptionPane.showMessageDialog(view, view.getTxtBuscar().getText());
-			this.myArticulo=this.myArticuloDao.buscarArticuloNombre(view.getTxtBuscar().getText());
-			
-			//JOptionPane.showMessageDialog(view, myArticulo);
-			if(myArticulo!=null){
-				view.getTxtArticulo().setText(myArticulo.getArticulo());
-				view.getTxtPrecio().setText("L. "+myArticulo.getPrecioVenta());
-				netBuscar=0;
-				netBuscar++;
-				
-			}
-			else{
-				myArticulo=null;
-				view.getTxtArticulo().setText("");
-				view.getTxtPrecio().setText("");
-			}
-		}
-		else{
-			myArticulo=null;
-			view.getTxtArticulo().setText("");
-			view.getTxtPrecio().setText("");
-		}
-		*/
 		
 		if(caracter=='+'){
 			if(filaPulsada>=0){
@@ -1738,35 +1706,6 @@ public void calcularTotales(){
 		
 		
 	}
-
-	/*
-	private void buscarMasOmenos(int p){
-		//se comprueba que hay algo que buscar
-				if(view.getTxtBuscar().getText().trim().length()!=0){
-					
-					myArticuloDao=new ArticuloDao(conexion);
-					//JOptionPane.showMessageDialog(view, "2");
-					//JOptionPane.showMessageDialog(view, view.getTxtBuscar().getText());
-					this.myArticulo=this.myArticuloDao.buscarArticuloNombre(view.getTxtBuscar().getText(),p);
-					
-					//JOptionPane.showMessageDialog(view, myArticulo);
-					if(myArticulo!=null){
-						view.getTxtArticulo().setText(myArticulo.getArticulo());
-						view.getTxtPrecio().setText("L. "+myArticulo.getPrecioVenta());
-						
-					}
-					else{
-						myArticulo=null;
-						view.getTxtArticulo().setText("");
-						view.getTxtPrecio().setText("");
-					}
-				}
-				else{
-					myArticulo=null;
-					view.getTxtArticulo().setText("");
-					view.getTxtPrecio().setText("");
-				}
-	}*/
 	
 	private void salir(){
 		this.view.setVisible(false);
@@ -1774,32 +1713,36 @@ public void calcularTotales(){
 		
 	}
 	private void guardar(){
-		
-		
+
 		if(view.getModeloTabla().getRowCount()>1){
 
+			setFactura();
+			boolean resulVendedor=false;
 
-					setFactura();
-					myFactura.setCodigoCaja(ConexionStatic.getUsuarioLogin().getCajaActiva().getCodigo());
+			ViewCargarVenderor viewVendedor=new ViewCargarVenderor(view);
+			CtlCargarVendedor ctlVendedor=new CtlCargarVendedor(viewVendedor);
 
-					boolean resultado=facturaOrdenesDao.registrar(myFactura);
+			resulVendedor=ctlVendedor.cargarVendedor();
+			if(resulVendedor) {
+				myFactura.setCodigoCaja(ConexionStatic.getUsuarioLogin().getCajaActiva().getCodigo());
+				myFactura.setVendedor(ctlVendedor.getVendedor());//activas para cuando se necesite un vendedor
 
-					if(resultado){
-						myFactura.setIdFactura(facturaDao.getIdFacturaGuardada());
-						resultado=true;
+				boolean resultado = facturaOrdenesDao.registrar(myFactura);
 
-						this.tipoView=1;
-						//this.view.setVisible(false);
-						//view.addBotonPendiente(myFactura,this);
+				if (resultado) {
+					myFactura.setIdFactura(facturaDao.getIdFacturaGuardada());
+					resultado = true;
+					this.tipoView = 1;
+					setEmptyView();
 
-						setEmptyView();
+					view.getBtnsGuardador().deleteAll();
+					myFactura.resetTotales();
 
-						view.getBtnsGuardador().deleteAll();
-
-						cargarFacturasPendientes(facturaOrdenesDao.facturasEnProceso());
-					}else{
-						JOptionPane.showMessageDialog(view, "Error al guardar la factura temporal", "Error al guardar", JOptionPane.ERROR_MESSAGE);
-					}
+					cargarFacturasPendientes(facturaOrdenesDao.facturasEnProceso());
+				} else {
+					JOptionPane.showMessageDialog(view, "Error al guardar la factura temporal", "Error al guardar", JOptionPane.ERROR_MESSAGE);
+				}
+			}
 
 		}else{
 			JOptionPane.showMessageDialog(view, "Para guardar debe agregar articulos.","ERROR",JOptionPane.ERROR_MESSAGE);
